remove sonnet dependency
